refactor: rename IoTooltip > IoTable component
It is used not only for tooltips
